APERTA-6893 Rework figure placement implementation
- Make a single pass through the document collecting possible captions.
- Match the candidates with existing figures in order.

The new method uses a regex rather than a finite list of delimiters, and
allows us to do the candidate processing in ruby land instead of in
nokogiri queries.

We removed a few tests that had been made redundant by the tests to
'call', as well as those for the private methods we removed.

@Bestra and @bnbeckwith and @diiq

APERTA-6893 Slightly rework figure insertion code