Update circle.yml configuration to use a stricter file glob syntax.
There is no way using standard file globbing to negate
entire words so we are ignoring all folders inside a spec/
directory that start with "f".

If you try to do [!features] or [^features] it will treat it
as character set and rather than doing what you hope it will
end up not matching on any directories that start with any of
those letters.

If you try to split it up as CircleCI's support suggested
like this, [^f][^e][^a], then it will fail on things where then
second letter was an "e" or the third letter an "a". This
results in not running all specs.

A fallback to this would be to auto-generate/update the circleyml and specifically list out all of the directories we want to look in. This would protect us in the case where we added new spec directories that started with f.