Refactoring: Use #rescind controller action/end-point rather than #destroy.
Semantically, rescind and destroy are two different things. If we ever want to hook into rescinding it should be separate from destroy. A destroy could happen for reasons that have nothing to do with an editor or staff intending to rescind.

P.S. this removes the InvitationsController#destroy end-point as there is nothing that currently needs to call that

P.P.S. Thank you @diiq for the upcoming test/pattern for rescind on invitation.js model. Made my life much easier.

APERTA-6984