Fix yet another flaky spec
Journals on the admin page were being sorted by name. When you changed
the name, it would sort as you typed. This caused the capybara tests to
lose focus.

Instead, let’s sort by id so the newest ones come first and the journals
do not move around as you change the name.