Update Emberize.class_name(...) helper to return the class name, not the base class.
This resolves an issue where pusher events were sending the front-end the base-class name (like task or attachment) and the update wasn't being applied successfully since ember needs to know the actual front-end model to match it up with (like supporting-information-task or adhoc-attachement).

Prior to this, if you uploaded an attachment on an adhoc task it was not showing the attachment without a browser refresh. 

APERTA-6881