Pulling out resizeTextarea into an inline function. Also use minHeight for setting the height of the textarea.
* add a 1px height offset otherwise IE11 still wants to scroll. :(

Note: there is a scrollHeight bug in 10/11 where accessing scrollHeight causes IE to re-render and this leads to flickering/disappearing content. The textarea works fine until the content causes the user to scroll. This issue is not limited to textareas and exists prior to the changes for this JIRA issue.

More information on the IE bug (we are not using this plugin, just useful conversation about IE issue): https://github.com/noraesae/perfect-scrollbar/issues/160

APERTA-6805