update store_dir path for uploaders
We've changed how the uploaders are mounted (and thus where data should be stored on S3) but we're not ready to move the data to new locations.
This change to store_dir is intended to be a stop-gap solution until we tackle versioning attachments next.

Paired with Zach Dennis (@zdennis)
APERTA-6881