APERTA-7103 Fix polymorphic array types.
The code was naively checking only the type for the first entry in an
array payload to see if the payload was polymorphic.  For instance, if
the payload request  a 'task', and got back ['otherTask', 'task'] that
would be polymorphic, but ['task', 'otherTask'] would stop after seeing
that the first record was 'task'.