Updating SnapshotService.snapshot_paper! to snapshot attachments explicitly
This intentionally does not take advantage of making a "paper.attachments" association because then we're relying on an implementation detail to know that we really want to snapshot figures, supporting information files, adhoc attachments, and question attachments.

So, add associations on paper to let us write explicit and clear code in the SnapshotService.

The goal of these changes is to make it easier to determine if an attachment has been snapshotted so we know whether or not to remove the S3 file. We're getting there closer.

APERTA-6881