Consolidate the various "attachment" snapshot serializers into a single Snapshot::Serializer
This chooses consistency over storage as this will snapshot potentially meaningless attachment values for certain kinds of attachments, but it means all of the attachments are treated identically.

This should be okay for diff'ing in the browser because there are front-end diffing components that choose what properties to diff/display based on what it is (e.g. a figure, supporting information, etc).

APERTA-6881