APERTA-6415 correctly serialize
- Use `:root` instead of `:key`
- Make sure SensitiveInformationUserSerializer's information is a
superset of the plain UserSerializer to avoid any future surprises
- The evaluation order of the serializer's class macros matters when it
tries to sideload objects with under the same key.  In this case, if the
:creator gets sideloaded under :users after that same user has already
been sideloaded by :collaborations then it becomes a no-op.  Running the
creator has_many first ensures that all the info will be present.  This
is really why I made sure the sensitive info is purely a superset of the
non-sensitive stuff