This resets to the primary key sequence on the attachments table.
This is necessary since postgresql does not do this on its own and all of the attachments/s3-migration work will INSERT a whole bunch of data into the attachments table.

So the attachments table will have IDs up to a high value (like 7001) but the primary key sequence will return a very low number (like 101) for its nextval. This will cause  unnecessary unique violation constraints.

If we end up with a maximum ID of 7001 then this migration ensures the nextval will be 7002.

APERTA-6881