APERTA-7246 Using `acts_as_paranoid` gem on nested_questions
...and nested_question_answers.

This will add a `deleted_at` timestamp to those models and prevent
them from being deleted.

The ruby gems version of `acts_as_paranoid` has a bug that prevented
this from working.  So I forced the github version and revision to
make sure we're not bitten by that.

The error was:

```
16:34:34 worker.1  | Gem Load Error is: undefined method
`validate_find_options' for class `Class'
16:34:34 worker.1  | Backtrace for gem load error is:
16:34:34 worker.1  |
/Users/mercieca/.rbenv/versions/2.2.3/lib/ruby/gems/2.2.0/gems/acts_as_paranoid-0.3.1/lib/acts_as_paranoid.rb:42:in
`alias_method'
```