APERTA-6719 Do not bother to remove revise task on rescind
If a decision is rescinded, we do not want to remove the old revise task
except in the case that the decision being rescinded is the first
major/minor revision decision. Even in that case, keeping a revise task
around is not a real problem.

Rather than fix this code for all the other use cases that are not
implemented now - e.g. we don't want to delete the revise task in the
case where the paper is in the 2nd or greater round of revision - we
decided to remove the code that deletes the revise task.