APERTA-5588 remove invitation details endpoint
The new UI for invitations requires us to show a truncated version of
the invitation body.  We decided to serialize the bodies with all
invitations rather than relying on a specific request.  Invitations are
not currently sideloaded by the generic `TaskSerializer` (which is used
by the TasksController index action, so we feel there's no risk of
sending down too much data.