Do not mark attachment has done until all steps involved in downloading are complete.
This is to avoid other parts of the system being notified of the attachment update, and  thinking it is done downloading before it's fully realized/usable.

This is to prevent the bug where trying to request a Figure from the client-side was trying to generate URL when the resource token wasn't in existence yet (but the attachment was marked as "done").

APERTA-7415