Adding a /api/manuscript_attachments/:id end-point to prevent Internal Server 500 errors
There are times when pusher/slanger will tell the the front-end about a created or updated ManuscriptAttachment. Previously this was trying to load the attachment from the /api/attachments/:id end-points, but that end-point only knows how to serve AdhocAttachment(s), so it was error'ing with an Internal Server 500 error because it was trying to due an R&P check for the associated task. But ManuscriptAttachment(s) don't relate to tasks, they relate to the paper.

So this change-set makes a new end-point that is responsible for serving ManuscriptAttachment(s).

Related change, don't return src, detail_src, or preview_src if an Attachment is not a public_resource. This prevents the AttachmentSerializer from blowing up when trying to serialize a ManuscriptAttachment to be sent to the browser.

Paired with @bnbeckwith

APERTA-7248