Paper Tracker queries, "VERSION DATE > ..." and "SUBMISSION DATE > ...", no longer include the date being searched.
Prior to this change-set the greater than search was still using the beginning of day boundary marker. This meant that "> 2016-04-12" queries would return results on April 12th (since the fields being queried internally are date/time fields and April 12th 12:01am is greater than April 12th, 12:00am). Intuitively, that represents a >= query. 

With this changeset the following rules apply:

  * > will return all papers submitted after date, not including the given date

APERTA-7365