Paper Tracker was incorrectly communicating how many days ago its version date or submission date was.
It was communicating an off-by error in the case where the date in question happened later in the day then whatever the current time is. It would treat this has happening a day earlier.

The relative-time implementation has been updated to always start from the end of "today". This is to ensure that anything that happened over the course of yesterday will be considered as "1 day ago". 

This treats the _whole_ day as a day rather than just counting 24-hour cycles from whatever the current time is.

APERTA-7365