Spike: Etching in a Site Admin role that has a wildcard permission.
The wildcard is intended to give them all permissions. This approach allows site admins to short-circuit additional permission queries.

We could remove the short-circuit logic and explicitly assign all of the permissions in the system to the Site Admin role, but given the scope of access this may produce unnecessary queries. This option seems simpler for the time being and more performant.

Paired with Zach Dennis (@zdennis)

APERTA-7504