APERTA-7414 clarify upload button placement
Previously the button would hide when an upload was active.  In certain
situations it looks really confusing; invitation uploads have the
save/cancel button for an invitation right next to an upload progress
bar so it looks like the cancel/save is related to the pending upload.
This change disables the upload button and keeps it around.