Adding inline class doc to the System model and removing unused methods.
The unused methods are not required because the System model is only used for the Site Admin role and the Site Admin role short-circuits authorization lookup. Because it short-circuits auth lookup it does not need to rely on associations as defined in config/initializers/z_authorizations.rb -- at least not at this time.

APERTA-7504