APERTA-7450 create 502.html page
The design assumes that rails is not available to serve assets. The page
inlines its own styles (ugh) and reaches out to only a minimum of local
resources—all of which should still be available if rails is not
running.

I attempted to conform to the mockup rather than styling in the log-in
page.