Install mail_safe gem on all environments ...
It seems the way we turn mail_safe on and off is by setting
MAILSAFE_REPLACEMENT_ADDRESS. When that var is set, we have an
initializer that attempts to require mail_safe. This was bailing on
capistrano deploys to vagrant because mail_safe wasn’t installed there.

It seems like it makes sense to install the gem, but trust that it
won’t be used unless MAILSAFE_REPLACEMENT_ADDRESS is set.