SiteAdmin's no longer get access to queries that require participations because they are no longer considered participants.
However, if a user who is a Site Admin is assigned to an object in the system (e.g. a Paper id=99) and that assignment has a role who marks them as a participant, then that will still be returned.

This is to avoid the overhead of a Site Admin loading every single object in the system because they access. For example, the dashboard will load all of the papers a user is participating in. Prior to this change, a user with the Site Admin role would load _every_ paper in the system which was wrong.

APERTA-7504