Update authorizations STI behavior so that being assigned to a descendant will properly look up the assigned_to_type using the model's base_class.
Consider the following hierarchy:

   SpecializedTask < Task < ActiveRecord::Base

When a user is assigned to a SpecializedTask we want assigned_to_type to query against SpecializedTask.base_class.name which is Task.name. This is due to how ActiveRecord stores type information. It stores the most general type on foreign records and the more specific type on the tasks record.

APERTA-7505