devDependencies to dependencies to simplify shrinkwrap process
https://docs.npmjs.com/cli/shrinkwrap#other-notes

For our application there are no other differences
between dependencies and devDependencies, because
no one ever is going to `npm install tahi`.
https://docs.npmjs.com/files/package.json#devdependencies
DevDependencies have nothing to do with
"dependencies that only run in development" like
in the bundler sense.

Taking out devDependencies makes it so that future
devs don't have to think about either the shrinkwrap
process or "do I put this in devDependencies or
dependencies?"