APERTA-7883 remove global file drop behavior
Our s3-file-uploader and file-uploader components both exhibit some very
scary behavior at the moment, mostly when a user drops a file into
APERTA:

- I wasn't properly destroying the file uploader, and there was a chance
the callbacks for a 'destroyed' uploader could fire when a user dropped
a file into the browser.

- More scary: We didn't set proper drop targets for uploaders, so when a
user drops a file anywhere in the browser, every uploader on the page
would add that file.

This commit fixes both problems.  It still keeps the browser's default
drop behavior for files, (for chrome, if you drop an image into the
browser it will open it up in that window), but now a user has to drop
the file on a valid file uploader to actually use it.