Only clean up initialized uploaders
For some reason a few instances of the jquery file uploader components
weren't fully initialized, despite running through what looks like a
synchronous call to the file uploader.  I changed the teardown method to
only run if the appropriate data is present on the DOM element.  Hacky
but pragmatic.

I also removed the `afterRender` scheduling for setting up the plugin,
as there's only 1 instance of it and there's no danger didInsertElement
is going to get called more than once.