Inline ember-cli-lazyloader
I'd like to start looking at using [yarn](https://yarnpkg.com/) as a
npm client. A fresh npm install takes 2 minutes and 40 seconds on my
machine, whereas a fresh yarn install takes about 40 seconds.
Our ember-cli-lazyloader dependency currently isn't
compatible.
This is the only package we have that's not listed on the npm registry.
It's a standalone repo for a small file, so I've brought the code inline
here.