APERTA-7697 Refining diffing
Editors weren't using the manuscript diff because of the noise
that was generated.

In this PR when a manuscript is diffed, we'll refine the diff
one step further.  This is done by looking for adjcent chunks
where one was added and the next removed.  These chunks are
then diffed at the word level to reduce the noise in the final
diff.