Add role hints to journal task types
The role hints provide a hint to downstream consumers what role we
intend this task to be for. This will be used to group the tasks on the
display side.

In practice, the word describing the `old_role` for these has become the
`role_hint`.

This run also went through a rubocop autocorrect run to resolve any
issues.

APERTA-7457