switch to flat early_posting key-value pair
previously we were generating
early_article_posting

"early_article_posting": {
  "consent": true,
},

but since we're no longer including the statement, this is unnecessary.
it's also going to be much harder to return this with consent: false
when there's no data to serialize.  making it a flat key/value pair
solves both problems.