APERTA-7633 disambiguate permissions in journal_factory_spec
This took me a while to nail down. Theses specs were originally failing because
there were two permissions records in the system for edit on AdHocForReviwersTask.
The only difference between the two was that one had more liberal states than
the other. When we were attempting to find the permission record, both records
would match since they both had the required states, but only one was correct.
The find_by would wind up returning the wrong one most of the time
(the one with extra states assigned) and the spec would fail.