Adding an index on tasks.title
This is to ensure that the post migrate task that ensures all tasks have the correct titles can be run.

On ci.aperta.tech there is no index and over a hundred thousand tasks. This is taken more than the allotted query time to complete since it is doing a full table scan.