Resolving flaky journal_factory_spec.
The reason this was flaky is that the AdHocForEditorsTask and AdHocForReviewersTask test was looking up the permissions outside the context of the role or states and then trying to check the role or states later. 

The problem with this is that two roles may have the same permission.action/applies_to but different states. When this happens the success of test is dependent on non-deterministic DB ordering of results.

I imagine the rest of the spec will need to be updated to follow a similar convention even though it's not problems, but they are not currently suffering from this issue AFAIK.