Add in PDF viewer
The `pdf.js-viewer` bundles up the PDF.js library with the functionality
of the reference viewer. The bundled pdf build and worker are bundled up
in the asset pipeline. These javascript components are lazy loaded only
on the manuscript view.

The default `viewer.html` is largely copied into the
`pdf-manuscript/template.hbs`.

* Add `pdf.js-viewer` in vendor
* Update asset pipeline to include viewer css/javascript
* Create new pdf-manuscript component
  * Handle lazy-loading of javascript files
  * Provide basic viewer html

APERTA-7828