APERTA-7810 record migration progress before AC change
After talking to PO we've decided that backfilling existing report
numbers for current manuscripts will be jarring; most users have
manually numbered them already and this will conflict. The updated AC
specify that we should only autonumber reviewers for manuscripts created
after this PR is merged.  I'll probably end up adding a new column to
the papers table to act as a flag