APERTA-8647 add capybara spec
Due to a quirk in the test environment we found that the qunit test we
wrote to check the observer's behavior was always passing.  We added a
capybara test on the reviewer report task to make sure we had a proper
failing test before continuing.

Rather than using the `wait_for_ajax` method we renamed and reused the
`wait_for_upload` method, which takes a sentinel proc.