Resolve flaky spec for journal factory.
The flakiness was due to anonymous classes that were created in memory. If the spec/models/invitation_spec.rb ran before the spec/services/journal_factory_spec.rb then the journal factory spec would have failures.

It is not super common for applications to depend on in-memory inheritance hierarchy to do things, but Aperta is not common in this regard as it uses the in-memory ::Task hierarchy to do lots of things (like assign permissions).

This means that any time a new task class (anonymous or otherwise) is added in the test environment it will automatically be picked up by a number of places in Aperta.

One of those places is the JournalFactory which creates permissions tied to specific tasks. This works fine except for anonymous classes (which have no name).

The fix here to is add a helper method for the JournalFactory spec that strips out anonymous classes.

By anonymous class I am referring to using Class.new syntax like:

    Class.new(Task)

There may be other ways to approach this and if someone puts forth a new solution (or convention) that bypasses this issue altogether.

Until then, may your tests stay green.

APERTA-8159