Force paper update notification for PDF
When the data extraction finishes, PDFs don't actually update the paper
model with any new data. This means that we need to force the
"paper:updated" notification with a small update to paper.

In the paper model, the body gets updated to "PDF Upload: <TIMESTAMP>".
This will have the benefit of triggering the right notifications and
putting something useful on paper.

APERTA-7828