Forcefully resolve an intermittent issue with the rake task: db:import_remote
Ensure that there is no connection to the database since we're about to drop and recreate it.

@ceautery and I were getting when running rake db:import_remote[ci]:

    dropdb: database removal failed: ERROR:  database "DB_NAME_HERE" is   being accessed by other users
    DETAIL:  There is 1 other session using the database.
    rake aborted!
     Error dropping and creating blank database. Is DB_NAME_HERE in use?

We ran git bisect to discover why it was connected, but we got different results each time we ran thru. 

This is not related to APERTA-8450 but was an obstacle when trying to verify that the migrations on that branch ran successfully against CI and prod backups.

APERTA-8450