Add a work around rake task for running the attachment report and emailing the aperta dev team.
This is because the current version of chruby deployed to SFO and SOMA environments has a bug in processing shell arguments and it results in improperly escape square brackets.

When chruby gets updated this task can go away and the config/schedule.rb can be updated to pass along the email via the CLI rake task invocation.

APERTA-8159