Update analyze_attachment_failures_report_spec.rb to run all of the examples using noon today.
Otherwise timezone offsets may switch the underlying day and cause the report to produce different results than if they were on the same day.

E.g. if it's 11:00 PM -0500 in UTC it's going to be the next day at 4:00 AM.

Note: this is to resolve spec failures from running this spec near day boundaries in various timezones.