Add .nvmrc file to project
This file specifies the node version that we use in production.
We do not use nvm in production, this is merely a convenience
to help developers that choose use use nvm in their local environment.

The goal here is to help devs create a consistent dev environment.

If you already have nvm installed, the command `nvm use` from anywhere
in this project should cause nvm to use the correct node version.

nvm does not automatically invoke `nvm use` upon entering the project.
If you would like that behavior, check out avn.