Add back in the notifications for failing uploads
An internal update was changed to one that did not push out
notifications. In the `DownloadAttachmentWorker` there is a section
where we handle logging of errors.

In this section we also set the attachment status to 'error' in a way
that will push out a notification to the clients.

APERTA-8853
APERTA-8485