Completely remove including HTML with epub files
There is no longer a use case for including HTML in posts to IHAT. This
commit removes the “include_html” flag, and uses only the behaviors that
were defined when include_html was false.

A few tests were affected by this, so their specs have been modified to
remove the concept of interrogating the content and publishing info
HTML files.

APERTA-7962