Remove tests referring to epubs with no source
There is no longer a use case (so I’ve been assured) of creating an epub
file without a source doc/docx attachment. This removes (failing) tests
that made that assumption, and the last of the tests that interrogate
the extracted HTML, which is no longer included in the epub file.

APERTA-7962