Hide questions unless we have a report
Unless there is an accepted invitation, there is no report to show. This
commit hides any questions to avoid confusion. This leaves just the
status text showing if the reviewer has been invited, declined, or if
the invitation has been rescinded.

This change also applies to past reports. Only the status text will be
shown.

APERTA-7993