Find or create reviewer reports
There is a possibility that a reviewer could be invited before a paper
is submitted. An acceptance of the invitation would create a reviewer
report for the current revision.

A submission of the paper would try to create the same report again
causing an exception. This guards against the case although it should
not be hit in production, it is possible in development.

APERTA-7993