Backfill paper ids on answers for reviewer report tasks
For the sake of consistency Answers have a presence validation on
paper_id.  Existing NestedQuestionAnswers have paper_ids in all cases
except those that are owned by ReviewerReports.  This commit adds the
`paper` method to the ReviewerReport class to allow those ids to be set
in the future, and fills in the paper_id column for the existing models.