Add “replace” buttons to upload manuscript card
This adds the S3 info into the Ember store for manuscript upload task
and source file upload. After a manuscript or source file is uploaded,
the upload button is swapped for the filename, which is a link to the
file just uploaded.

This also fixes a breaking test for sending to Apex, which created a
manuscript without a pending_url field, which is now required for
rails manuscript parsing to work.