Add fake Cards to data dump
Due to the way that development has worked, there needs to be some fake
Cards loaded into the system so that the Admin Cards screen is not 
blank.  This commit adds a rake task to get them in the system and
updates the data.yml so that it gets loaded automatically for QA.

Some (very soon) future work will be superseding all of this and once
real Cards are in the system in a reliable way, these changes will be
safe to blow away.