Don't blow up if an Answerable doesn't have a card
For some existing tasks, like
TahiStandardTasks::SupportingInformationTask we aren't currently
creating a blank card.  In that case, `Task.card_for` could blow up.  In
the short term I'm loosening the restriction until we decide if we want
to create blank cards for all the existing tasks, not just the ones that
have nested questions.