Card seed rake tasks become Configuration classes
Previously, there existed a rake task for every Card that needed to be
created in the system.  This commit moves each of these from a rake 
task into a class responsible for defining the necessary properties of
the Card.  Something else will be responsible for loading this
configuration class and actually creating the Card.