Disable reviewer status
Review status are broken, but most of the core functionality is merged.
Instead of pulling it out, this PR disables the user-facing view of the
status.

Note that I kept the output for tests so I wouldn't have to
disable any existing tests. This is done by adding a `testing` property
to the component. This property should be removed in the final fix.

APERTA-7993