Adjust error icon positioning
The alert icon was not always vertically centered in the input text field.
This was due to whether the caller wanted to show a label or change the
label sizing.  Instead, group the input and the notification icon
separately from the input and the adjust the position relative to that
grouping.