Add CardVersion model
CardVersions sit between a card and the various versions of its
content.  card_versions is a join table between cards and
card_contents.  This allows us to keep the concept of identity with the
Card model while keeping track of the stuff that changes over time
with new CardVersions.  Cards no longer have a single piece of
root content, but rather a root for each version.