Keep reviewer reports up to date with invitations
Add AASM to ReviewerReports to track three states for before an accepted
invitation, after accept and after submission.

Add notifications to the states on invitation (and reorder some updates)
so that the right report is updated as the invitation state changes.

And update ReviewerReport to better handle these new cases.

APERTA-7993