replace vendored code with bowered code
and provide a symlink to bowered pdf.js-viewer

instead of having a rake task to manage our vendored code with
modifications, we've got our fork of pdf.js-viewer in a git repo
and can install it with bower (the recommended installation
method by the maintainers of the upstream repo).

This should be much easier to manage in the long run, and will
allow us to submit changes upstream.  (They've already accepted
one very small bug fix.)