Show field level error messages
Allow creation of either text inputs or textareas.  This should be
enhanced and extracted later when there are additional form input
examples.  Currently, there is only a need for these two.

Instead of showing all errors on all fields in the form, enhance the
component to show only the errors specific to each field.  If a field
happens to have multiple error messages, concatenate the message before
displaying it.