Make adhoc attachment snapshots available to Ember
This tells the revise-task component to download snapshot data for the
current paper, and makes adhoc attachments available to the view by
filtering on owner_id. The decision-bar component then filters the
adhoc attachments by the decision’s revision number. (The filenames are
not yet displayed in the view.)

APERTA-7039