changes manually applied
I corrected what appears to be a mistake in the previous changes.
This line was removed before rather than changed:

    PDFJS.workerSrc = pdfjsSrc.replace(/pdf\.js$/i, 'pdf.worker.js');

PDFJS.workerSrc is used later in the code, so I changed the line in
a manner consistent with the other changes, rather than remove the line.

I also changed another instance of the 'thumbnail' CSS class to
'pdfthumbnail', since it seems necessary that we change them all
if we are going to maintain the previous functioning of PDF.js.