Drop status, status_datetime, add completed_at
And compute the rest. Thanks to the state machine, we no longer need to
store the `status` and `status_datetime` in the database. Depending on
the state, we can find the status on the invitation or mark it complete
and store that timestamp in `completed_at`.

This makes the serializer compute the status on the fly with the
provided methods.

APERTA-7993