Background the journal logo processing
Previously, the journal logo processing was happening inline.  So,
improve performance by backgrounding this and leaning on journal event
streaming to return the processed logo when complete.

In addition, allow the logo url to be sent with the update and create
parameters rather than a separate call.  Retire the upload_logo route.