Check sidekiq-redis instead of just redis service
The redis service will be checked as part of the sidekiq-redis check
since it is a dependency.  In addition, there is already a custom
health check in this file that confirms the ability to write to the
currrent redis instance.

In addition, in the current version of `health_check` gem, there is no
way to configure the redis instance that should be verified.  It
assumes that it will always be localhost which is not the case on heroku
since it is using a cloud based redis provider.  Although the
`health_check` gem has the `redis` gem as a dependency and *that* gem
looks at the `REDIS_URL` environment variable to see what redis instance
to use, the `REDIS_URL` variable is not automatically set by the cloud
based redis provider Aperta currently uses on the pipelined review
apps.  So, use a custom redis connection url when provided.