Introduce CustomCardTask
This is a descendant of the normal Task model, but will be the only one
that is associated to CardContent (through Card -> CardVersion).

This model is being used to make incremental changes to Aperta as
custom card content is being introduced.  The majority of the app
revolves around Tasks, so this model will make cards quack like one.