Use update_column for data migration
When running the data migration in the same process that is handling the
addition of a new column, it seems to not save data to the column
(silently failing). By using `update_column` we can force the data into
the database.

A new assertion at the end checks to ensure that any 'submitted'
reviews have a `submitted_at` datetime set.

APERTA-7993