Refresh decision’s attachments after upload (WIP)
Reload /decisions after a new decision_attachment is uploaded on the
Response to Reviewers card. This is sub-optimal, and should be replaced
eventually with a refresh of latestRegisteredDecision, but some more
Ember wiring is needed to get that working.

This also adds a Rails route to allow get requests for individual
decision_attachments.

APERTA-7039