Migrate the heroku database upon each release
We are currently using the heroku pipeline functionality which will 
establish a new deployment app for each PR we create.  When a deployment
app is first created, it will run the "postdeploy" hook in the app.json
file.  Currently, this is configured to run `db:setup`

Unfortunately, this "postdeploy" hook only runs once -- when the
deployment heroku app is first created.  This becomes a real problem
when someone provides feedback on an existing PR, updates and pushes it,
but because the "postdeploy" hook doesn't run, any new database changes
are never seen.

This PR fixes this by taking advantage of a new heroku beta feature
called release phase* which gives us a hook that runs every time a 
deployment happens.

* https://devcenter.heroku.com/articles/release-phase