fix extremely minor bug which changes error message returned by health-check
instead of our message, the bug caused the health-check default
to be returned.  The only practical effect of this bug is the
potential to confuse future developers if it is not fixed,
as the behavior is not 100% what is expected.

it does not prevent the health check from working, though.