APERTA-9738 change div[contenteditable] style
Currently contenteditable divs are display: list-item to fix a firefox
rendering bug as part of APERTA-9418.  Unfortunately that renders a list dot
next to each contenteditable.  We can either use list-style: none, or go with
the approach that I've picked here, simply adding a min height to the
contenteditable to fix the specific problem.