Experiment with ditching pdfjs-viewer
The pdfjs docs (https://mozilla.github.io/pdf.js/getting_started/)
suggest that it’s inappropriate to use the pdf js viewer embedded in
a custom application. I think that if we drop down a layer and just
use the pdfs “display” api we could solve a few problems. We’d be 
implementing our own “viewer” which we could customize for better visual
 and technical compatibility with our application.

This would mean that we’d have to recreate some of the functionality 
that pdfs-viewer currently provides. But I think we’ve customized the 
viewer to the point, and run into enough hassles with upgrading it,
 that totally owning the “viewer” layer would lead to a less complicated system to maintain.

Examples in the pdfjs docs show rudimentary examples of what we could do
https://mozilla.github.io/pdf.js/examples/