Select radio button option regardless of datatype
When determining if the radio button should be selected or not,
coerce both the html radio button form element value and any
current answer to strings.  This is to protect against the case
where the answered value in the database is being returned as
a non-string datatype (e.g., `true` instead of "true").  At
the html level, the value will always be a string, so do a
string-to-string comparison to ensure that the radio button
is properly selected or not.