Update shrinkwrap and package.json for APERTA-9045
This adds font-awesome 4.7 to package.json, which was needed to handle a
missing dependency after a clean npm install. The resulting shrinkwrap
file contains several non-breaking version increments of main modules,
and a handful of breaking changes isolated to dependencies of main
modules. A quick smoke test didn’t reveal any side-effects from this.

One exception to the above was broccoli-funnel, which needed to be
upgraded from 0.2.3 to 1.1.0 to work correctly with ember-font-awesome.
Of the many packages that retained the same version number, a few later
in the shrinkwrap file aren’t grouped correctly by git-diff, so they
look like they have been deleted and re-added, e.g., ember-cli and
ember-can.

The main packages all fall within the range specified in package.json,
so I suspect our master shrinkwrap file has, like myself, needed a
haircut for quite a while. Due to the sheer number of changes here, I’ll
ask QA to take special care to check for anything out of the ordinary.