Put billing log spec into "maintenance mode"
Each example in this spec takes almost a second on my new macbook pro, and none
of them are doing anything earthshattering.  It simply takes a while to create
the required models.  Rather than trying to share the setup I collapsed the
expectations that have shared setups into a single example under an
aggregate_failures block.  That way they run faster but preserve their
granularity if we need to diagnose future failures.