Introduce method for finding Cards by class name.
When creating new answerables (i.e., Task), an existing Card needs to
be associated to it.  The only way to know which Card to use for many
existing answerables is looking up the Cards using the class name of the
answerable itself.

The actual Card is created using the configuration files found in 
/lib/tasks/card_loading/configurations

Since this logic is needed in many places, it is being wrapped up and
treated as a convenience method here in the Card model.