Update all answerables with Card associations
To be valid, all answerables will need an association to a Card via the
CardVersion.  So, ensure all factories are providing valid answerables.

In addition, move all tahi_standard_task task factories to the main app.
Reasoning for this is to that all tasks can leverage the :with_card
trait and prevent code duplication.  In addition, we need to stop
pretending that rails engines in this project are siloed containers and
instead embrace the fact that code and setup bleed throughout the engines.